requirements.txt
# Core dependencies
aiohttp==3.10.10
asyncio==3.4.3
python-dotenv==1.0.1
pydantic==2.9.2

# Database
aiosqlite==0.20.0

# Firecrawl client
httpx==0.27.2
tenacity==9.0.0

# RSS parsing
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0

# Media processing
Pillow==10.4.0
playwright==1.48.0
aiofiles==24.1.0

# Monitoring and API
fastapi==0.115.4
uvicorn[standard]==0.32.0
websockets==13.1

# Logging
python-json-logger==2.0.7
colorama==0.4.6

# Utilities
python-dateutil==2.9.0
pytz==2024.2
humanize==4.11.0

# WordPress integration
openai==1.56.2
requests==2.32.3
transliterate==1.10.2